apply more merging...
