+++ Domain.Contracts/Repositories/IModuleRepository.cs
@@ -4,5 +4,7 @@ namespace Domain.Contracts.Repositories;
 
 public interface IModuleRepository : IRepositoryBase<Module>
 {
-
+    Task<(IEnumerable<Module> Modules, int TotalCount)> GetAllModulesAsync(
+        int page, int pageSize, int? courseId, bool trackChanges = false);
+    Task<Module?> GetModuleAsync(int id, bool trackChanges = false);
 }
